+++ detectree2/models/train.py
@@ -1,3 +1,8 @@
+"""Train a model.
+
+Classes and functions to train a model based on othomosaics and corresponding
+manual crown data.
+"""
 import datetime
 import glob
 import json
@@ -254,6 +259,7 @@ class MyTrainer(DefaultTrainer):
                 is_train=True,
                 augmentations=[
                     T.Resize((800, 800)),    # is this necessary/helpful?
+                    # T.ResizeScale
                     T.RandomBrightness(0.8, 1.8),
                     T.RandomContrast(0.6, 1.3),
                     T.RandomSaturation(0.8, 1.4),
@@ -381,8 +387,13 @@ def combine_dicts(root_dir: str,
         for d in train_dirs:
             tree_dicts += get_tree_dicts(d)
         return tree_dicts
-    else:
+    elif mode == "val":
         tree_dicts = get_tree_dicts(train_dirs[(val_dir - 1)])
+        return tree_dicts
+    elif mode == "full":
+        tree_dicts = []
+        for d in train_dirs:
+            tree_dicts += get_tree_dicts(d)
         return tree_dicts
 
 
@@ -405,12 +416,26 @@ def get_filenames(directory: str):
     return dataset_dicts
 
 
-def register_train_data(train_location, name="tree", val_fold=1):
-    for d in ["train", "val"]:
+def register_train_data(train_location, name: str = "tree", val_fold=None):
+    """Register data for training and (optionally) validation
+
+    Args:
+        train_location: directory containing training folds
+        name: string to name data
+        val_fold: fold assigned for validation and tuning. If not given,
+        will take place on all folds.
+    """
+    if val_fold is not None:
+        for d in ["train", "val"]:
+            DatasetCatalog.register(
+                name + "_" + d,
+                lambda d=d: combine_dicts(train_location, val_fold, d))
+            MetadataCatalog.get(name + "_" + d).set(thing_classes=["tree"])
+    else:
         DatasetCatalog.register(
-            name + "_" + d,
-            lambda d=d: combine_dicts(train_location, val_fold, d))
-        MetadataCatalog.get(name + "_" + d).set(thing_classes=["tree"])
+            name + "_" + "full",
+            lambda d=d: combine_dicts(train_location, 0, "full"))
+        MetadataCatalog.get(name + "_" + "full").set(thing_classes=["tree"])
 
 
 def remove_registered_data(name="tree"):
@@ -451,13 +476,13 @@ def setup_cfg(base_model:
               num_classes=1,
               eval_period=100,
               out_dir="/content/drive/Shareddrives/detectree2/train_outputs"):
-    """Set up config object
+    """Set up config object.
 
     Args:
-        base_model:
-        trains:
-        tests:
-        update_model:
+        base_model: base pre-trained model from detectron2 model_zoo
+        trains: names of registered data to use for training
+        tests: names of registered data to use for evaluating models
+        update_model: updated pre-trained model from detectree2 model_garden
         workers:
         ims_per_batch:
         gamma:
@@ -476,13 +501,13 @@ def setup_cfg(base_model:
     cfg.DATASETS.TRAIN = trains
     cfg.DATASETS.TEST = tests
     cfg.DATALOADER.NUM_WORKERS = workers
-    #cfg.SOLVER.IMS_PER_BATCH = ims_per_batch
+    cfg.SOLVER.IMS_PER_BATCH = ims_per_batch
     cfg.SOLVER.GAMMA = gamma
     cfg.MODEL.BACKBONE.FREEZE_AT = backbone_freeze
     cfg.SOLVER.WARMUP_ITERS = warm_iter
     cfg.SOLVER.MOMENTUM = momentum
-    #cfg.MODEL.RPN.BATCH_SIZE_PER_IMAGE = batch_size_per_im
-    #cfg.SOLVER.WEIGHT_DECAY = 0.001
+    cfg.MODEL.RPN.BATCH_SIZE_PER_IMAGE = batch_size_per_im
+    cfg.SOLVER.WEIGHT_DECAY = 0.001
     cfg.SOLVER.BASE_LR = base_lr
     cfg.OUTPUT_DIR = out_dir
     os.makedirs(cfg.OUTPUT_DIR, exist_ok=True)
@@ -506,8 +531,7 @@ def predictions_on_data(directory=None,
                         scale=1,
                         geos_exist=True,
                         num_predictions=0):
-    """
-    Prediction produced from a test folder and outputted to predictions folder
+    """Prediction produced on a test folder
     """
 
     test_location = directory + "test"
