+++ docs/source/conf.py
@@ -12,7 +12,9 @@
 #
 import os
 import sys
+
 import sphinx_rtd_theme
+
 sys.path.insert(0, os.path.abspath('../detectree2'))
